feature-compress-incremental/requirements.txt
torch>=2.6
torchvision>=0.17
transformers>=4.44
pycocotools>=2.0
tqdm>=4.65
datasets>=4.2.0
numpy>=1.24
Pillow>=10.0
scikit-learn>=1.2
